requirements.txt
openai-whisper
pydub
ffmpeg